refactoring infer serializers, fix tests
