+++ src/serializers/mod.rs
@@ -16,6 +16,7 @@ mod config;
 mod errors;
 mod extra;
 mod filter;
+mod infer;
 mod ob_type;
 mod shared;
 mod type_serializers;
