+++ src/serializers/type_serializers/datetime_etc.rs
@@ -6,8 +6,10 @@ use pyo3::types::{PyDate, PyDateTime, PyDict, PyTime};
 use crate::build_context::BuildContext;
 use crate::input::{pydate_as_date, pydatetime_as_datetime, pytime_as_time};
 
-use super::any::{fallback_json_key, fallback_serialize, fallback_to_python};
-use super::{py_err_se_err, BuildSerializer, CombinedSerializer, Extra, SerMode, TypeSerializer};
+use super::{
+    infer_json_key, infer_serialize, infer_to_python, py_err_se_err, BuildSerializer, CombinedSerializer, Extra,
+    SerMode, TypeSerializer,
+};
 
 pub(crate) fn datetime_to_string(py_dt: &PyDateTime) -> PyResult<String> {
     let dt = pydatetime_as_datetime(py_dt)?;
@@ -63,7 +65,7 @@ macro_rules! build_serializer {
                         extra
                             .warnings
                             .on_fallback_py(self.get_name(), value, error_on_fallback)?;
-                        fallback_to_python(value, include, exclude, extra)
+                        infer_to_python(value, include, exclude, extra)
                     }
                 }
             }
@@ -80,7 +82,7 @@ macro_rules! build_serializer {
                         extra
                             .warnings
                             .on_fallback_py(self.get_name(), key, error_on_fallback)?;
-                        fallback_json_key(key, extra)
+                        infer_json_key(key, extra)
                     }
                 }
             }
@@ -103,7 +105,7 @@ macro_rules! build_serializer {
                         extra
                             .warnings
                             .on_fallback_ser::<S>(self.get_name(), value, error_on_fallback)?;
-                        fallback_serialize(value, serializer, include, exclude, extra)
+                        infer_serialize(value, serializer, include, exclude, extra)
                     }
                 }
             }
