+++ src/serializers/type_serializers/mod.rs
@@ -25,6 +25,11 @@ pub mod with_default;
 pub(self) use super::config::utf8_py_error;
 pub(self) use super::errors::{py_err_se_err, PydanticSerializationError};
 pub(self) use super::extra::{Extra, ExtraOwned, SerMode};
-pub(self) use super::filter::{AnyFilter, SchemaFilter};
+pub(self) use super::filter::SchemaFilter;
+pub(self) use super::infer::{
+    infer_json_key, infer_serialize, infer_serialize_known, infer_to_python, infer_to_python_known, SerializeInfer,
+};
 pub(self) use super::ob_type::{IsType, ObType};
-pub(self) use super::shared::{to_json_bytes, BuildSerializer, CombinedSerializer, PydanticSerializer, TypeSerializer};
+pub(self) use super::shared::{
+    object_to_dict, to_json_bytes, BuildSerializer, CombinedSerializer, PydanticSerializer, TypeSerializer,
+};
