+++ src/serializers/type_serializers/typed_dict.rs
@@ -9,10 +9,10 @@ use serde::ser::SerializeMap;
 use crate::build_context::BuildContext;
 use crate::build_tools::{py_error_type, schema_or_config, SchemaDict};
 
-use super::any::{fallback_json_key, fallback_serialize, fallback_to_python, SerializeInfer};
 use super::with_default::get_default;
 use super::{
-    py_err_se_err, BuildSerializer, CombinedSerializer, Extra, PydanticSerializer, SchemaFilter, TypeSerializer,
+    infer_json_key, infer_serialize, infer_to_python, py_err_se_err, BuildSerializer, CombinedSerializer, Extra,
+    PydanticSerializer, SchemaFilter, SerializeInfer, TypeSerializer,
 };
 
 #[derive(Debug, Clone)]
@@ -170,7 +170,7 @@ impl TypeSerializer for TypedDictSerializer {
                             }
                         }
                         if self.include_extra {
-                            let value = fallback_to_python(value, include, exclude, extra)?;
+                            let value = infer_to_python(value, include, exclude, extra)?;
                             new_dict.set_item(key, value)?;
                         }
                     }
@@ -181,7 +181,7 @@ impl TypeSerializer for TypedDictSerializer {
                 extra
                     .warnings
                     .on_fallback_py(self.get_name(), value, error_on_fallback)?;
-                fallback_to_python(value, include, exclude, extra)
+                infer_to_python(value, include, exclude, extra)
             }
         }
     }
@@ -232,7 +232,7 @@ impl TypeSerializer for TypedDictSerializer {
                         }
                         if self.include_extra {
                             let s = SerializeInfer::new(value, include, exclude, extra);
-                            let output_key = fallback_json_key(key, extra).map_err(py_err_se_err)?;
+                            let output_key = infer_json_key(key, extra).map_err(py_err_se_err)?;
                             map.serialize_entry(&output_key, &s)?
                         }
                     }
@@ -243,7 +243,7 @@ impl TypeSerializer for TypedDictSerializer {
                 extra
                     .warnings
                     .on_fallback_ser::<S>(self.get_name(), value, error_on_fallback)?;
-                fallback_serialize(value, serializer, include, exclude, extra)
+                infer_serialize(value, serializer, include, exclude, extra)
             }
         }
     }
