+++ src/serializers/type_serializers/url.rs
@@ -6,8 +6,10 @@ use pyo3::types::PyDict;
 use crate::build_context::BuildContext;
 use crate::url::{PyMultiHostUrl, PyUrl};
 
-use super::any::{fallback_json_key, fallback_serialize, fallback_to_python};
-use super::{BuildSerializer, CombinedSerializer, Extra, SerMode, TypeSerializer};
+use super::{
+    infer_json_key, infer_serialize, infer_to_python, BuildSerializer, CombinedSerializer, Extra, SerMode,
+    TypeSerializer,
+};
 
 macro_rules! build_serializer {
     ($struct_name:ident, $expected_type:literal, $extract:ty) => {
@@ -45,7 +47,7 @@ macro_rules! build_serializer {
                         extra
                             .warnings
                             .on_fallback_py(self.get_name(), value, error_on_fallback)?;
-                        fallback_to_python(value, include, exclude, extra)
+                        infer_to_python(value, include, exclude, extra)
                     }
                 }
             }
@@ -62,7 +64,7 @@ macro_rules! build_serializer {
                         extra
                             .warnings
                             .on_fallback_py(self.get_name(), key, error_on_fallback)?;
-                        fallback_json_key(key, extra)
+                        infer_json_key(key, extra)
                     }
                 }
             }
@@ -82,7 +84,7 @@ macro_rules! build_serializer {
                         extra
                             .warnings
                             .on_fallback_ser::<S>(self.get_name(), value, error_on_fallback)?;
-                        fallback_serialize(value, serializer, include, exclude, extra)
+                        infer_serialize(value, serializer, include, exclude, extra)
                     }
                 }
             }
